refactor: separate vesting from timelock
