+++ apps/wallet-dashboard/lib/interfaces/index.ts
@@ -2,4 +2,5 @@
 // SPDX-License-Identifier: Apache-2.0
 
 export * from './transactions.interface';
+export * from './timelocks.interface';
 export * from './vesting.interface';
