+++ apps/wallet-dashboard/lib/interfaces/vesting.interface.ts
@@ -1,42 +1,6 @@
 // Copyright (c) 2024 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 
-export interface UID {
-    id: string;
-}
-
-export interface Balance {
-    value: number;
-}
-
-export interface Timelocked {
-    id: UID;
-    locked: Balance;
-    expirationTimestampMs: number;
-    label?: string;
-}
-
-export interface TimelockedIotaResponse {
-    id: UID;
-    locked: string;
-    expiration_timestamp_ms: string;
-    label?: string;
-}
-
-export interface StakedIota {
-    id: UID;
-    poolId: string;
-    stakeActivationEpoch: number;
-    principal: Balance;
-}
-
-export interface TimelockedStakedIota {
-    id: UID;
-    stakedIota: StakedIota;
-    expirationTimestampMs: number;
-    label?: string | null | undefined;
-}
-
 export enum SupplyIncreaseUserType {
     Staker = 'Staker',
     Entity = 'Entity',
