+++ apps/wallet-dashboard/lib/utils/timelock.ts
@@ -1,21 +1,18 @@
 // Copyright (c) 2024 IOTA Stiftung
 // SPDX-License-Identifier: Apache-2.0
 
-import { SUPPLY_INCREASE_VESTING_LABEL } from '../constants';
-import { Timelocked, TimelockedStakedIota } from '../interfaces';
+import { TimelockedObject, TimelockedStakedIota } from '../interfaces';
 
 export function isTimelockedStakedIota(
-    obj: Timelocked | TimelockedStakedIota,
+    obj: TimelockedObject | TimelockedStakedIota,
 ): obj is TimelockedStakedIota {
     const referenceProperty: keyof TimelockedStakedIota = 'stakedIota';
     return referenceProperty in obj;
 }
 
-export function isTimelocked(obj: Timelocked | TimelockedStakedIota): obj is Timelocked {
-    const referenceProperty: keyof Timelocked = 'locked';
+export function isTimelockedObject(
+    obj: TimelockedObject | TimelockedStakedIota,
+): obj is TimelockedObject {
+    const referenceProperty: keyof TimelockedObject = 'locked';
     return referenceProperty in obj;
-}
-
-export function isVesting(obj: Timelocked | TimelockedStakedIota): boolean {
-    return obj.label === SUPPLY_INCREASE_VESTING_LABEL;
 }
