+++ apps/wallet-dashboard/lib/utils/vesting/vesting.ts
@@ -14,17 +14,18 @@ import {
     SupplyIncreaseUserType,
     SupplyIncreaseVestingPayout,
     SupplyIncreaseVestingPortfolio,
-    Timelocked,
     TimelockedIotaResponse,
+    TimelockedObject,
     TimelockedStakedIota,
     VestingOverview,
 } from '../../interfaces';
-import { isTimelocked, isTimelockedStakedIota, isVesting } from '../timelock';
+import { isTimelockedObject, isTimelockedStakedIota } from '../timelock';
+import { isSupplyIncreaseVestingObject } from '../vesting';
 
 export function getLastSupplyIncreaseVestingPayout(
-    objects: (Timelocked | TimelockedStakedIota)[],
+    objects: (TimelockedObject | TimelockedStakedIota)[],
 ): SupplyIncreaseVestingPayout | undefined {
-    const vestingObjects = objects.filter(isVesting);
+    const vestingObjects = objects.filter(isSupplyIncreaseVestingObject);
 
     if (vestingObjects.length === 0) {
         return undefined;
@@ -38,14 +39,14 @@ export function getLastSupplyIncreaseVestingPayout(
 }
 
 function supplyIncreaseVestingObjectsToPayoutMap(
-    vestingObjects: (Timelocked | TimelockedStakedIota)[],
+    vestingObjects: (TimelockedObject | TimelockedStakedIota)[],
 ): Map<number, SupplyIncreaseVestingPayout> {
     const expirationToVestingPayout = new Map<number, SupplyIncreaseVestingPayout>();
 
     for (const vestingObject of vestingObjects) {
         let objectValue = 0;
-        if (isTimelocked(vestingObject)) {
-            objectValue = (vestingObject as Timelocked).locked.value;
+        if (isTimelockedObject(vestingObject)) {
+            objectValue = (vestingObject as TimelockedObject).locked.value;
         } else if (isTimelockedStakedIota(vestingObject)) {
             objectValue = (vestingObject as TimelockedStakedIota).stakedIota.principal.value;
         }
@@ -111,10 +112,10 @@ export function buildSupplyIncreaseVestingSchedule(
 }
 
 export function getVestingOverview(
-    objects: (Timelocked | TimelockedStakedIota)[],
+    objects: (TimelockedObject | TimelockedStakedIota)[],
     currentEpochTimestamp: number,
 ): VestingOverview {
-    const vestingObjects = objects.filter(isVesting);
+    const vestingObjects = objects.filter(isSupplyIncreaseVestingObject);
     const latestPayout = getLastSupplyIncreaseVestingPayout(vestingObjects);
 
     if (vestingObjects.length === 0 || !latestPayout) {
@@ -148,7 +149,7 @@ export function getVestingOverview(
         0,
     );
 
-    const timelockedObjects = vestingObjects.filter(isTimelocked);
+    const timelockedObjects = vestingObjects.filter(isTimelockedObject);
 
     const totalAvailableClaimingAmount = timelockedObjects.reduce(
         (acc, current) =>
@@ -185,7 +186,7 @@ export function getSupplyIncreaseVestingPayoutsCount(userType: SupplyIncreaseUse
     return SUPPLY_INCREASE_VESTING_PAYOUTS_IN_1_YEAR * vestingDuration;
 }
 
-export function mapTimelockObjects(iotaObjects: IotaObjectData[]): Timelocked[] {
+export function mapTimelockObjects(iotaObjects: IotaObjectData[]): TimelockedObject[] {
     return iotaObjects.map((iotaObject) => {
         if (!iotaObject?.content?.dataType || iotaObject.content.dataType !== 'moveObject') {
             return {
